use largestTxFee for bond
