+++ src/modals/UpdateBond/Forms/BondAll.tsx
@@ -7,23 +7,21 @@ import { useApi } from 'contexts/Api';
 import { useConnect } from 'contexts/Connect';
 import { useModal } from 'contexts/Modal';
 import { useActivePools } from 'contexts/Pools/ActivePools';
-import { useTheme } from 'contexts/Themes';
 import { useTransferOptions } from 'contexts/TransferOptions';
 import { useTxFees } from 'contexts/TxFees';
 import { EstimatedTxFee } from 'library/EstimatedTxFee';
 import { Warning } from 'library/Form/Warning';
+import useBondGreatestFee from 'library/Hooks/useBondGreatestFee';
 import { useSubmitExtrinsic } from 'library/Hooks/useSubmitExtrinsic';
 import { useEffect, useState } from 'react';
-import { defaultThemes } from 'theme/default';
 import { planckBnToUnit } from 'Utils';
-import { NotesWrapper, Separator } from '../../Wrappers';
+import { Separator } from '../../Wrappers';
 import { FormsProps } from '../types';
 import { FormFooter } from './FormFooter';
 
 export const BondAll = (props: FormsProps) => {
   const { setSection, setLocalResize } = props;
 
-  const { mode } = useTheme();
   const { api, network } = useApi();
   const { units } = network;
   const { setStatus: setModalStatus, config } = useModal();
@@ -32,6 +30,7 @@ export const BondAll = (props: FormsProps) => {
   const { bondType } = config;
   const { txFees, txFeesValid } = useTxFees();
   const { selectedActivePool } = useActivePools();
+  const largestTxFee = useBondGreatestFee({ bondType });
 
   let { unclaimedRewards } = selectedActivePool || {};
   unclaimedRewards = unclaimedRewards ?? new BN(0);
@@ -53,11 +52,11 @@ export const BondAll = (props: FormsProps) => {
   const [bond, setBond] = useState({ bond: freeBalance });
 
   // bond minus tx fees
-  const bondAfterTxFees = BN.max(freeBalanceBn.sub(txFees), BN_ZERO);
+  const bondAfterTxFees = BN.max(freeBalanceBn.sub(largestTxFee), BN_ZERO);
 
   // total possible bond after tx fees
   const totalPossibleBond = planckBnToUnit(
-    BN.max(totalPossibleBondBn.sub(txFees), BN_ZERO),
+    BN.max(totalPossibleBondBn.sub(largestTxFee), BN_ZERO),
     units
   );
 
@@ -126,7 +125,9 @@ export const BondAll = (props: FormsProps) => {
           )}
           <h4>Amount to bond:</h4>
           <h2>
-            {planckBnToUnit(bondAfterTxFees, units)} {network.unit}
+            {largestTxFee.eq(new BN(0))
+              ? '...'
+              : `${planckBnToUnit(bondAfterTxFees, units)} ${network.unit}`}
           </h2>
           <p>
             This amount of {network.unit} will be added to your current bonded
@@ -135,16 +136,12 @@ export const BondAll = (props: FormsProps) => {
           <Separator />
           <h4>New total bond:</h4>
           <h2>
-            {totalPossibleBond} {network.unit}
+            {largestTxFee.eq(new BN(0))
+              ? '...'
+              : `${totalPossibleBond} ${network.unit}`}
           </h2>
-          <NotesWrapper>
-            {txFees.gt(BN_ZERO) && (
-              <p style={{ color: defaultThemes.text.success[mode] }}>
-                Transaction fees have been deducted from your total bond amount.
-              </p>
-            )}
-            <EstimatedTxFee />
-          </NotesWrapper>
+          <Separator />
+          <EstimatedTxFee />
         </>
       </div>
       <FormFooter
